requirements.txt
# fastapi as the main framework
fastapi==0.110.0
# pydantic for data validation & to handle config file
pydantic==2.6.3
pydantic-settings==2.2.1
pydantic_core==2.16.3
starlette==0.36.3
# uvicorn as the server
uvicorn==0.15.0
# to handle multipart form data
python-multipart==0.0.9
# to handle json web tokens
python-jose[cryptography]
python-jose==3.3.0
# to handle password hashing
passlib==1.7.4
# to handle database
SQLAlchemy==2.0.27
# A PostgreSQL adapter for Python
psycopg2-binary==2.9.9
# encrypting and decrypting strings
bcrypt==4.1.2

python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
# to handle http requests in testing
httpx==0.27.0
# to handle testing
pytest==8.0.2
# to handle database migrations
alembic==1.13.1
# to handle communication with modbus devices like PLCs
pyModbusTCP==0.2.1